fix(api): disable StripeModule in desktop mode
- Updated the `startApiServerForElectron` function to utilize `LoggerService` for improved logging consistency.
- Refactored the subscription handling logic by moving webhook handlers to a new `SubscriptionWebhooks` class, enhancing code organization and maintainability.
- Reintroduced Stripe webhook handling for checkout sessions and subscription events, ensuring proper processing and logging of subscription-related actions.
- Updated `package.json` build scripts to include necessary file copying for generated sources, improving build efficiency.
- Commented out legacy test cases in `subscription.service.spec.ts` for future reference and clarity.
